add StaticAndInstance Variable
