+++ WindowsFormsApp_practice/02_WindowsFormsApp_practice.csproj
@@ -59,7 +59,7 @@
       <DependentUpon>Form4.cs</DependentUpon>
     </Compile>
     <Compile Include="partialClass.cs" />
-    <Compile Include="MyClass.cs" />
+    <Compile Include="Utility.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
