+++ WindowsFormsApp_practice/partialClass.cs
@@ -7,7 +7,7 @@ using System.Windows.Forms;
 
 namespace WindowsFormsApp_practice
 {
-    internal partial class MyClass
+    internal partial class Utility
     // Point: Partail class要原本的與新增的class在同一個nameSpace下且class名一樣，並加上partial
     {
         public void method06()
